Created a classify module to abstract the score/grade classification.
trident.classify provides a base class for grade classification which
has an "interpolate" function. The argument of this function should be
the trident score dict.

trident.classify provides two template classification functions, one
that use the scipy linear interpolator and another that combines the
linear interpolator for multiple species, mapping the interpolator to
the species abbreviation in the 'query_id' section of the score.